Move vaults members endpoint
